# src 2/main/resources/openapi/app-api.yaml
openapi: 3.0.3
info:
  title: Order IT App API
  description: |-
    API for Order IT Application
  contact:
    email: info@mycloudpa.com
  version: 1.0.0
...
paths:
...
  /api/suppliers:
    get:
      tags:
        - supplier
      summary: Get all suppliers
      description: Get all suppliers without products
      operationId: getSuppliers
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                description: suppliers
                items:
                  $ref: '#/components/schemas/Supplier'
        '400':
          description: Invalid status value
  /api/suppliers/{supplierId}/products:
    get:
      tags:
        - supplier
      summary: get all products for a given supplier
      description: get all products for a given supplier
      operationId: getSupplierProductsBySupplierId
      parameters:
        - in: path
          name: supplierId
          schema:
            type: number
          required: true
          description: identifier of the Supplier
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupplierProduct'
        '400':
          description: Invalid status value
...
  /api/company/profile:
    put:
      tags:
        - company
      summary: update profile of a company
      description: update profile of a company
      operationId: updateCompanyProfile
      requestBody:
        description: company profile to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyProfile'
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
...
components:
  schemas:
...
    CompanyProfile:
      required:
        - contactNumber
      type: object
      properties:
        contactNumber:
          type: string
...
    Supplier:
      type: object
      properties:
        id:
          type: number
          description: identifier of the Supplier
        supplierName:
          type: string
          description: supplier name
        description:
          type: string
          description: supplier description
        cutOff:
          type: string
          description: date or time that the supplier stop taking new orders
        supplierLogoUrl:
          type: string
          description: logo url for the Supplier
        supplierEmail:
          type: string
          description: supplier email
        contactNumber:
          type: string
          description: supplier contact number
        addressLine1:
          type: string
          description: supplier address line1
        addressLine2:
          type: string
          description: supplier address line2
        addressLine3:
          type: string
          description: supplier address line3
        deliveryDay:
          type: string
          description: on what days the supplier make delivery
        minOrder:
          type: string
          description: min order amount
        category:
          type: string
          description: supplier category
        shopName:
          type: string
          description: supplier shopName
        accountNumber:
          type: string
          description: supplier accountNumber
        isIntegratedSupplier:
          type: boolean
          description: whether supplier is an integrated supplier or not
        supplierProducts:
          type: array
          description: products the supplier sell
          items:
            $ref: '#/components/schemas/SupplierProduct'
    SupplierProduct:
      type: object
      properties:
        id:
          type: number
          description: identifier of the Product
        supplierId:
          type: number
          description: identifier of the Product's Supplier
        productName:
          type: string
          description: product name
        description:
          type: string
          description: product description
        productImageUrl:
          type: string
          description: image url for the product
        nickname:
          type: string
          description: nickname of the product
        unit:
          type: string
          description: unit of the product
        vat:
          type: number
          description: VAT of the Product
        price:
          type: number
          description: unit price of the Product
...